slot in new client
