+++ libddog/crud/client.py
@@ -1,11 +1,14 @@
+import json
+import logging
 import os
-from typing import Any, List, Optional
+import time
+from typing import Dict, List, Optional, Type
 
-import datadog
-import datadog.api
+import requests
 
 from libddog.common.types import JsonDict
 from libddog.crud.errors import (
+    AbstractCrudError,
     DashboardCreateFailed,
     DashboardDeleteFailed,
     DashboardGetFailed,
@@ -15,11 +18,34 @@ from libddog.crud.errors import (
     MissingDatadogAppKey,
 )
 from libddog.dashboards import Dashboard
+from libddog.tools.logs import enable_logging
 
 
 class DatadogClient:
     env_varname_api_key = "DATADOG_API_KEY"
     env_varname_app_key = "DATADOG_APPLICATION_KEY"
+
+    def __init__(self) -> None:
+        self.api_key: Optional[str] = None
+        self.app_key: Optional[str] = None
+
+        self.baseurl = "https://api.datadoghq.com/api/v1"
+
+        self.session = requests.Session()
+
+        self.logger = logging.getLogger(__name__)
+
+        # Force enable logging because we want to see our own log messages. This
+        # is not the greatest solution; ideally this should be done at the
+        # entrypoint to the program. But we have different code paths that lead
+        # us here, so doing this purely in bin/ddog would not affect logging in
+        # tests and different tests may wish to consume the client at different
+        # layers. Doing this at the innermost layer where we know logging is
+        # definitely used is the easiest way to avoid missing code paths.
+        # We need to force enable because at this point it's likely that some
+        # other library has init'd logging already with a different set of
+        # parameters.
+        enable_logging(force=True)
 
     def load_credentials_from_environment(self) -> None:
         var_api_key = self.env_varname_api_key
@@ -36,55 +62,155 @@ class DatadogClient:
             error = f"Could not find {var_app_key!r} set in the environment"
             raise MissingDatadogAppKey(errors=[error])
 
-        options = {
-            "api_key": api_key,
-            "app_key": app_key,
-        }
-        datadog.initialize(**options)  # type: ignore
+        self.api_key = api_key
+        self.app_key = app_key
 
-    def parse_response_errors(self, resp: Any) -> List[str]:
-        if type(resp) == dict:
-            return resp.get("errors") or []
+    def prepare_headers(self) -> Dict[str, str]:
+        assert self.api_key is not None
+        assert self.app_key is not None
+
+        headers = {
+            "Content-Type": "application/json",
+            "DD-API-KEY": self.api_key,
+            "DD-APPLICATION-KEY": self.app_key,
+        }
+        return headers
+
+    def build_dashboard_url(self, id: Optional[str] = None) -> str:
+        url = f"{self.baseurl}/dashboard"
+
+        if id is not None:
+            url = f"{url}/{id}"
+
+        return url
+
+    def try_parse_json_payload(self, response: requests.Response) -> Optional[JsonDict]:
+        try:
+            payload: JsonDict = response.json()
+            return payload
+        except json.decoder.JSONDecodeError:
+            pass
+
+        return None
+
+    def try_parse_payload_errors(self, payload: Optional[JsonDict]) -> List[str]:
+        if isinstance(payload, dict):
+            return payload.get("errors") or []
 
         return []
+
+    def try_parse_ratelimit_reset(self, response: requests.Response) -> Optional[float]:
+        value = response.headers.get("X-RateLimit-Reset")
+
+        if value:
+            try:
+                return float(value)
+            except ValueError:
+                pass
+
+        return None
+
+    def make_request(
+        self,
+        *,
+        request: requests.Request,
+        expected_code: int,
+        exc_cls: Type[AbstractCrudError],
+    ) -> Optional[JsonDict]:
+        response: Optional[requests.Response] = None
+        payload: Optional[JsonDict] = None
+        errors: List[str] = []
+
+        attempt_no = 1
+        wait_secs_base = 0.5
+        prepared_request = request.prepare()
+
+        while attempt_no < 4:
+            try:
+                response = self.session.send(prepared_request)
+            except requests.exceptions.RequestException as exc:
+                errors.append(str(exc))
+
+            if response is not None and response.status_code == 429:
+                wait_secs = self.try_parse_ratelimit_reset(response)
+                if wait_secs is None:
+                    # there is no jitter here but I don't think it matters
+                    # because this is a fallback code path anyway
+                    wait_secs = wait_secs_base * (2 ** attempt_no)
+
+                # log a warning because we are deliberately pausing execution
+                self.logger.warning(
+                    f"Request was rate limited by the Datadog API, "
+                    f"retrying in {wait_secs} seconds"
+                )
+                time.sleep(wait_secs)
+                attempt_no += 1
+                continue
+
+            break
+
+        if response is not None:
+            payload = self.try_parse_json_payload(response)
+            errors = self.try_parse_payload_errors(payload) or []
+
+        if response is None or response.status_code != expected_code or errors:
+            status_code = response.status_code if response is not None else None
+            raise exc_cls(errors=errors, http_status_code=status_code)
+
+        return payload
 
     def create_dashboard(self, dashboard: Dashboard) -> str:
         client_kwargs = dashboard.as_dict()
         client_kwargs.pop("id", None)  # we cannot pass an id when creating
 
-        resp = datadog.api.Dashboard.create(**client_kwargs)  # type: ignore
-        errors = self.parse_response_errors(resp)
-        if errors:
-            raise DashboardCreateFailed(errors=errors)
+        url = self.build_dashboard_url()
+        headers = self.prepare_headers()
+        request = requests.Request(
+            method="POST", url=url, headers=headers, json=client_kwargs
+        )
 
-        id: str = resp["id"]
+        payload = self.make_request(
+            request=request, expected_code=200, exc_cls=DashboardCreateFailed
+        )
+
+        assert isinstance(payload, dict)  # help mypy
+        id = payload["id"]
+        assert isinstance(id, str)  # help mypy
+
         return id
 
     def delete_dashboard(self, *, id: str) -> None:
-        result = datadog.api.Dashboard.delete(id=id)  # type: ignore
+        url = self.build_dashboard_url(id=id)
+        headers = self.prepare_headers()
+        request = requests.Request(method="DELETE", url=url, headers=headers)
 
-        errors = self.parse_response_errors(result)
-        if errors:
-            raise DashboardDeleteFailed(errors=errors)
+        self.make_request(
+            request=request, expected_code=200, exc_cls=DashboardDeleteFailed
+        )
 
     def get_dashboard(self, *, id: str) -> JsonDict:
-        result = datadog.api.Dashboard.get(id=id)  # type: ignore
+        url = self.build_dashboard_url(id=id)
+        headers = self.prepare_headers()
+        request = requests.Request(method="GET", url=url, headers=headers)
 
-        errors = self.parse_response_errors(result)
-        if errors:
-            raise DashboardGetFailed(errors=errors)
+        payload = self.make_request(
+            request=request, expected_code=200, exc_cls=DashboardGetFailed
+        )
 
-        assert isinstance(result, dict)  # help mypy
-        return result
+        assert isinstance(payload, dict)  # help mypy
+        return payload
 
     def list_dashboards(self) -> List[JsonDict]:
-        result = datadog.api.Dashboard.get_all()  # type: ignore
+        url = self.build_dashboard_url()
+        headers = self.prepare_headers()
+        request = requests.Request(method="GET", url=url, headers=headers)
 
-        errors = self.parse_response_errors(result)
-        if errors:
-            raise DashboardListFailed(errors=errors)
+        payload = self.make_request(
+            request=request, expected_code=200, exc_cls=DashboardListFailed
+        )
 
-        dashboards = result["dashboards"]
+        assert isinstance(payload, dict)  # help mypy
+        dashboards = payload["dashboards"]
         assert isinstance(dashboards, list)  # help mypy
 
         return dashboards
@@ -100,7 +226,12 @@ class DatadogClient:
         client_kwargs = dashboard.as_dict()
         client_kwargs.pop("id", None)  # we pass it separately
 
-        resp = datadog.api.Dashboard.update(id=id, **client_kwargs)  # type: ignore
-        errors = self.parse_response_errors(resp)
-        if errors:
-            raise DashboardUpdateFailed(errors=errors)
+        url = self.build_dashboard_url(id=id)
+        headers = self.prepare_headers()
+        request = requests.Request(
+            method="PUT", url=url, headers=headers, json=client_kwargs
+        )
+
+        self.make_request(
+            request=request, expected_code=200, exc_cls=DashboardUpdateFailed
+        )
